fix swagger and run eslint
