+++ src/database/index.ts
@@ -2,7 +2,7 @@ import { AppAnonymousReportModel } from '@models/app/anonymousReport';
 import { AppMediaContentModel } from '@models/app/appMediaContent';
 import { AppMediaThumbnailModel } from '@models/app/appMediaThumbnail';
 import { AppNotificationModel } from '@models/app/appNotification';
-import { UbiBeneficiarySurveyModel } from '@models/ubi/ubiBeneficiarySurvey';
+import { AppProposalModel } from '@models/app/appProposal';
 import { AppUserModel } from '@models/app/appUser';
 import { AppUserThroughTrustModel } from '@models/app/appUserThroughTrust';
 import { AppUserTrustModel } from '@models/app/appUserTrust';
@@ -29,6 +29,7 @@ import { Inflow } from '@models/ubi/inflow';
 import { Manager } from '@models/ubi/manager';
 import { UbiRequestChangeParamsModel } from '@models/ubi/requestChangeParams';
 import { UbiBeneficiaryRegistryModel } from '@models/ubi/ubiBeneficiaryRegistry';
+import { UbiBeneficiarySurveyModel } from '@models/ubi/ubiBeneficiarySurvey';
 import { UbiBeneficiaryTransactionModel } from '@models/ubi/ubiBeneficiaryTransaction';
 import { UbiClaimModel } from '@models/ubi/ubiClaim';
 import { ClaimLocationModel } from '@models/ubi/ubiClaimLocation';
@@ -45,7 +46,6 @@ import { Sequelize, Options, ModelCtor } from 'sequelize';
 import config from '../config';
 import { DbModels } from '../types/db';
 import initModels from './models';
-import { AppProposalModel } from '@models/app/appProposal';
 
 let logging:
     | boolean
