+++ src/database/models/ubi/community.ts
@@ -1,90 +1,8 @@
-import { AppMediaContent } from '@interfaces/app/appMediaContent';
-import { AppProposal } from '@interfaces/app/appProposal';
-import { StoryCommunity } from '@interfaces/story/storyCommunity';
-import { UbiClaimLocation } from '@interfaces/ubi/ubiClaimLocation';
-import { UbiCommunityContract } from '@interfaces/ubi/ubiCommunityContract';
-import { UbiCommunityDailyMetrics } from '@interfaces/ubi/ubiCommunityDailyMetrics';
-import { UbiCommunityDailyState } from '@interfaces/ubi/ubiCommunityDailyState';
-import { UbiCommunityDemographics } from '@interfaces/ubi/ubiCommunityDemographics';
-import { UbiCommunityState } from '@interfaces/ubi/ubiCommunityState';
-import { UbiCommunitySuspect } from '@interfaces/ubi/ubiCommunitySuspect';
+import {
+    CommunityAttributes,
+    ICommunityCreationAttributes,
+} from '@interfaces/ubi/community';
 import { Sequelize, DataTypes, Model } from 'sequelize';
-
-import { ICommunityContractParams } from '../../../types';
-// import { UbiPromoter } from '@interfaces/ubi/ubiPromoter';
-
-import { BeneficiaryAttributes } from './beneficiary';
-
-export interface IBaseCommunityAttributes {
-    requestByAddress: string;
-    name: string;
-    contractAddress?: string | undefined;
-    description: string;
-    language: string;
-    currency: string;
-    city: string;
-    country: string;
-    gps: {
-        latitude: number;
-        longitude: number;
-    };
-    email: string;
-    coverMediaId?: number;
-    contractParams?: ICommunityContractParams;
-}
-
-export interface CommunityAttributes {
-    id: number; // Note that the `null assertion` `!` is required in strict mode.
-    publicId: string; // TODO: to be removed
-    requestByAddress: string;
-    contractAddress: string | null;
-    name: string;
-    description: string;
-    descriptionEn: string | null;
-    language: string;
-    currency: string;
-    city: string;
-    country: string;
-    gps: {
-        latitude: number;
-        longitude: number;
-    };
-    email: string;
-    visibility: 'public' | 'private';
-    coverImage: string; // TODO: to be removed
-    coverMediaId: number;
-    status: 'pending' | 'valid' | 'removed'; // pending / valid / removed
-    review: 'pending' | 'in-progress' | 'halted' | 'closed';
-    started: Date; // TODO: to be removed
-    proposalId: number | null;
-
-    // timestamps
-    createdAt: Date;
-    updatedAt: Date;
-    deletedAt: Date | null;
-
-    metrics?: UbiCommunityDailyMetrics[]; // TODO: to be removed
-    cover?: AppMediaContent;
-    proposal?: AppProposal;
-    contract?: UbiCommunityContract; // TODO: to be removed
-    state?: UbiCommunityState; // TODO: to be removed
-    storyCommunity?: StoryCommunity[]; // TODO: to be removed
-    suspect?: UbiCommunitySuspect[]; // TODO: to be removed
-    beneficiaries?: BeneficiaryAttributes[]; // TODO: to be removed
-    // promoter?: UbiPromoter;
-    claimLocation?: UbiClaimLocation[]; // TODO: to be removed
-    demographics?: UbiCommunityDemographics[]; // TODO: to be removed
-    dailyState?: UbiCommunityDailyState[]; // TODO: to be removed
-}
-export interface ICommunityCreationAttributes extends IBaseCommunityAttributes {
-    descriptionEn?: string;
-    visibility?: 'public' | 'private';
-    coverImage?: string; // TODO: will be required once next version is released
-    status?: 'pending' | 'valid' | 'removed'; // pending / valid / removed
-    started?: Date;
-    txReceipt?: any | undefined;
-    contractParams?: ICommunityContractParams;
-}
 
 export class Community extends Model<
     CommunityAttributes,
