+++ src/database/models/ubi/manager.ts
@@ -1,8 +1,7 @@
 import { AppUser } from '@interfaces/app/appUser';
+import { CommunityAttributes } from '@interfaces/ubi/community';
 import { UbiBeneficiaryRegistry } from '@interfaces/ubi/ubiBeneficiaryRegistry';
 import { Sequelize, DataTypes, Model } from 'sequelize';
-
-import { CommunityAttributes } from './community';
 
 export interface ManagerAttributes {
     id: number;
