+++ src/interfaces/app/appProposal.ts
@@ -1,3 +1,22 @@
+/**
+ * @swagger
+ *  components:
+ *    schemas:
+ *      AppProposal:
+ *        type: object
+ *        required:
+ *          - id
+ *          - status
+ *          - endBlock
+ *        properties:
+ *          id:
+ *            type: integer
+ *          status:
+ *            type: integer
+ *          endBlock:
+ *            type: integer
+ */
+
 export interface AppProposal {
     id: number;
     status: number;
