+++ src/interfaces/ubi/ubiCommunityDailyState.ts
@@ -1,3 +1,45 @@
+/**
+ * @swagger
+ *  components:
+ *    schemas:
+ *      UbiCommunityDailyState:
+ *        type: object
+ *        required:
+ *          - id
+ *          - communityId
+ *          - gps
+ *          - createdAt
+ *        properties:
+ *          id:
+ *            type: integer
+ *          communityId:
+ *            type: string
+ *          claimed
+ *            type: string
+ *          claims
+ *            type: integer
+ *          beneficiaries
+ *            type: integer
+ *          raised
+ *            type: string
+ *          backers
+ *            type: integer
+ *          monthlyBackers
+ *            type: integer
+ *          volume
+ *            type: string
+ *          transactions
+ *            type: number
+ *          reach
+ *            type: number
+ *          reachOut
+ *            type: number
+ *          fundingRate
+ *            type: number
+ *          date
+ *            type: string
+ */
+
 export interface UbiCommunityDailyState {
     id: number;
     communityId: number;
