+++ src/services/ubi/claimLocation.ts
@@ -1,8 +1,8 @@
-import { BeneficiaryAttributes } from '@models/ubi/beneficiary';
+import { BeneficiaryAttributes } from '@interfaces/ubi/beneficiary';
 import { BaseError } from '@utils/baseError';
-import config from '../../config';
 import { Op } from 'sequelize';
 
+import config from '../../config';
 import { models } from '../../database';
 
 export default class ClaimLocationService {
@@ -69,7 +69,9 @@ export default class ClaimLocationService {
         }[]
     > {
         const fiveMonthsAgo = new Date();
-        fiveMonthsAgo.setDate(fiveMonthsAgo.getDate() - config.claimLocationTimeframe);
+        fiveMonthsAgo.setDate(
+            fiveMonthsAgo.getDate() - config.claimLocationTimeframe
+        );
         return models.ubiClaimLocation.findAll({
             attributes: ['gps'],
             where: {
