+++ tests/factories/beneficiaryTransaction.ts
@@ -1,7 +1,7 @@
 import { ethers } from 'ethers';
 
-import { BeneficiaryAttributes } from '../../src/database/models/ubi/beneficiary';
 import { UbiBeneficiaryTransactionModel } from '../../src/database/models/ubi/ubiBeneficiaryTransaction';
+import { BeneficiaryAttributes } from '../../src/interfaces/ubi/beneficiary';
 import { UbiBeneficiaryTransactionCreation } from '../../src/interfaces/ubi/ubiBeneficiaryTransaction';
 import { randomTx } from '../utils/utils';
