+++ tests/integration/services/story.test.ts
@@ -3,8 +3,8 @@ import { Op, Sequelize } from 'sequelize';
 import Sinon, { assert, replace, spy, match } from 'sinon';
 
 import { models } from '../../../src/database';
-import { CommunityAttributes } from '../../../src/database/models/ubi/community';
 import { AppUser } from '../../../src/interfaces/app/appUser';
+import { CommunityAttributes } from '../../../src/interfaces/ubi/community';
 import { StoryContentStorage } from '../../../src/services/storage';
 import StoryService from '../../../src/services/story';
 import BeneficiaryFactory from '../../factories/beneficiary';
